requirements.txt
hydra-core
uproot
awkward-array
mplhep